fix another compile issue

+++ src/common/comm/client.cc
@@ -39,7 +39,7 @@ int main(int argc, char *argv[]) {
 	const LTFSDmProtocol::LTFSDmMigRequestResp migreqresp = command.migrequestresp();
 
 	if( migreqresp.success() == true ) {
-		printf("sending success, token: %lu\n", migreqresp.token());
+		printf("sending success, token: %llu\n", (unsigned long long) migreqresp.token());
 		if ( getpid() != migreqresp.pid() )
 			printf("WRONG RESPONSE -- WRONG RESPONSE -- WRONG RESPONSE -- WRONG RESPONSE\n");
 	}
